refactor: Move codes into structure
