+++ src/utilities/dateUtils.js
@@ -1,0 +1,12 @@
+const toLocaleDateString = (date) => {
+    return date.toLocaleDateString('en-US', {
+        day: 'numeric',
+        month: 'long',
+        year: 'numeric',
+    });
+}
+
+
+module.exports = {
+    toLocaleDateString
+}
